# openapi.yaml
openapi: 3.0.3

info:
  title: LLM Backend
  description: |
    Thin wrapper around an LLM provider (currently Anthropic) used by
    distributed coding agents in the CS6650 cache-coherence project.

    Two levels of API are exposed:

    - **`/task`** — structured agent endpoint. Accepts a repo, a list of
      context files, the file to modify, and the task description. The backend
      assembles the prompt and passes it to the LLM.
    - **`/generate`** — raw LLM call. Useful for baselines and direct
      experiments where prompt assembly is handled by the caller.

    `KVState` is an opaque map returned alongside every response. It is empty
    while the backend uses Anthropic (the API handles prefix caching
    server-side). When the backend switches to llama.cpp, `KVState` will carry
    the key/value cache tensors for the shared context prefix so agents can
    skip reprocessing tokens they have already seen.
  version: 0.1.0

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: agent
    description: High-level agent task endpoint
  - name: llm
    description: Low-level LLM generation endpoint
  - name: observability
    description: Health and metrics

paths:

  /task:
    post:
      tags: [agent]
      summary: Submit a structured coding task
      description: |
        Builds a prompt from the supplied repository context and task
        description, calls the LLM, and returns the generated response.

        The backend assembles the prompt in this order:
        1. Repository header (repo + branch)
        2. Contents of each `context_files` entry (read-only reference material)
        3. Contents of `target_file` (the file to be modified)
        4. The `task` instruction

        **KVState note:** once the backend moves to llama.cpp, the shared
        prefix formed by `context_files` will be cached and reused across
        agents working on the same repo, reducing redundant token processing.
        A file-level lock on `target_file` will be added to prevent conflicts
        between concurrent agents.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
            examples:
              basic:
                summary: Single context file
                value:
                  repo: "owner/my-repo"
                  context_files: ["src/utils.py"]
                  target_file: "src/main.py"
                  task: "Add input validation to the parse() function"
              multi_context:
                summary: Multiple context files, non-default branch
                value:
                  repo: "owner/my-repo"
                  context_files: ["src/models.py", "src/utils.py"]
                  target_file: "src/api.py"
                  task: "Add a DELETE /item/{id} endpoint"
                  branch: "feature/api-v2"
                  max_tokens: 2048
      responses:
        '202':
          description: |
            Task accepted. Processing runs asynchronously after this response
            is returned — the caller does not wait for the LLM or commit to
            complete. Failures are logged server-side. Poll `/metrics` to
            observe aggregate throughput, or check CloudWatch logs for
            per-request errors.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAccepted'
              example:
                request_id: "a3f1c2d4-89ab-4e12-b456-426614174000"
                status: "accepted"
        '422':
          $ref: '#/components/responses/ValidationError'

  /generate:
    post:
      tags: [llm]
      summary: Raw LLM generation
      description: |
        Sends a pre-assembled prompt directly to the LLM. No context fetching
        or prompt building is performed. Use this for baselines, ablations, or
        when the caller manages prompt construction itself.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
            examples:
              simple:
                summary: Plain prompt
                value:
                  prompt: "Explain the difference between strong and eventual consistency."
              with_system:
                summary: Prompt with system instruction
                value:
                  prompt: "Refactor this function to use a context manager."
                  system: "You are an expert Python developer. Reply with only the updated code."
                  max_tokens: 512
      responses:
        '200':
          description: LLM generation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '502':
          $ref: '#/components/responses/UpstreamError'

  /health:
    get:
      tags: [observability]
      summary: Liveness check
      responses:
        '200':
          description: Service is running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok

  /metrics:
    get:
      tags: [observability]
      summary: Aggregate request metrics
      description: |
        Returns cumulative counters since the last process start. Metrics reset
        on container restart — persistent storage (CloudWatch / Prometheus) is
        handled at the infrastructure layer.
      responses:
        '200':
          description: Aggregate metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsResponse'

components:
  schemas:

    TaskAccepted:
      type: object
      required: [request_id, status]
      properties:
        request_id:
          type: string
          format: uuid
          description: Unique ID for this task. Use it to correlate log entries.
        status:
          type: string
          enum: [accepted]

    TaskRequest:
      type: object
      required: [repo, context_files, target_file, task]
      properties:
        repo:
          type: string
          description: Repository identifier in `owner/repo` format.
          example: "owner/my-repo"
        context_files:
          type: array
          items:
            type: string
          description: |
            Paths to files the agent needs as read-only context. Order matters:
            files listed first appear earlier in the prompt and are more likely
            to be reused as a cached prefix across agents working on the same
            repo.
          example: ["src/models.py", "src/utils.py"]
        target_file:
          type: string
          description: Path to the file the agent should modify.
          example: "src/main.py"
        task:
          type: string
          description: Natural-language description of the change to make.
          example: "Add input validation to the parse() function"
        branch:
          type: string
          description: Branch to read files from.
          default: "main"
        max_tokens:
          type: integer
          description: Maximum tokens in the LLM response.
          default: 1024
          minimum: 1

    GenerateRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          description: The fully-assembled prompt to send to the LLM.
        max_tokens:
          type: integer
          default: 1024
          minimum: 1
        system:
          type: string
          nullable: true
          description: Optional system prompt. Null means no system prompt is sent.

    GenerateResponse:
      type: object
      required: [content, input_tokens, output_tokens, latency_ms]
      properties:
        content:
          type: string
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        latency_ms:
          type: number
          format: float

    MetricsResponse:
      type: object
      required:
        - total_requests
        - total_input_tokens
        - total_output_tokens
        - total_latency_ms
        - avg_latency_ms
      properties:
        total_requests:
          type: integer
          description: Total number of LLM calls since last restart.
        total_input_tokens:
          type: integer
          description: Cumulative input tokens across all requests.
        total_output_tokens:
          type: integer
          description: Cumulative output tokens across all requests.
        total_latency_ms:
          type: number
          format: float
          description: Cumulative LLM call time in milliseconds.
        avg_latency_ms:
          type: number
          format: float
          description: Mean LLM call latency in milliseconds.

    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]

    ErrorDetail:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Human-readable error description.

  responses:
    ValidationError:
      description: Request body failed schema validation.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items: {}
                    msg:
                      type: string
                    type:
                      type: string

    UpstreamError:
      description: The upstream LLM API returned an error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Anthropic API error: 529 overloaded"